Add rule to filter files by references
Update the _NodeMapFileWidgetState to include a new rule that filters files based on their references. This enables the user to view files that have references to the selected file. The fileAnalyzer object is now used to update the rule and pass the filtered files to the MapMindBasePage. This change improves the file exploration capabilities of the NodeMapFileWidget.