Add layer filtering functionality to FilesAnalyzerInfo
This commit adds the ability to filter layers based on a given condition in the FilesAnalyzerInfo class. The `layerByFilter` method now accepts an optional `filter` parameter and returns a filtered list of layers based on that condition. This allows for more granular control over the layers that are returned. This change enhances the flexibility and extensibility of the FilesAnalyzerInfo class.